chore: add jest and github action
We need to be able to ensure that the functionality is correct so we
make it possible to run `yarn jest` to validate.
