+++ package.json
@@ -14,6 +14,10 @@
     "@posthog/plugin-scaffold": "^1.3.4"
   },
   "devDependencies": {
-    "@types/node": "^18.11.18"
+    "@types/jest": "^29.5.2",
+    "@types/node": "^18.11.18",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.1.6"
   }
 }
